+++ src/components/MattressList/MattressList.styled.js
@@ -1,5 +1,6 @@
 import styled from "styled-components";
 import { BreadWrapper } from "../BreadCrumbs";
+import { colors, fontSize, fonts, boxShadow } from "../../utils/styles";
 
 export const NewBread = styled(BreadWrapper)`
   margin-left: 0;
@@ -41,12 +42,36 @@ export const MattListWrapper = styled.div`
   .mattList__grid {
     display: grid;
     grid-template-columns: repeat(auto-fill, minmax(265px, 1fr));
-    grid-auto-rows: minmax(300px, auto);
+    grid-auto-rows: minmax(300px, 427px);
     grid-gap: 1rem;
     margin-left: 7px;
     margin-right: 7px;
     @media screen and (min-width: 768px) {
       margin: 0;
+      width: 70%;
+    }
+    @media screen and (min-width: 1024px) {
+      width: 80%;
+    }
+  }
+  .noFilter {
+    background-color: ${colors.white};
+    box-shadow: ${boxShadow.md};
+    font-family: ${fonts.sans};
+    margin-left: 7px;
+    margin-right: 7px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    h3 {
+      border-bottom: 4px solid ${colors.red[800]};
+    }
+    border-bottom: 4px solid ${colors.red[800]};
+    @media screen and (min-width: 768px) {
+      margin: 0;
+      width: 70%;
+    }
+    @media screen and (min-width: 1024px) {
       width: 80%;
     }
   }