+++ src/components/shared/ProductList/Header.jsx
@@ -1,13 +1,28 @@
 import React from "react";
-import { Link } from "gatsby";
+// import { Link } from "gatsby";
 import styled from "styled-components";
 import PropTypes from "prop-types";
-import { colors } from "../../utils/styles";
+import { colors, spacing, boxShadow } from "../../../utils/styles";
+import { PrimaryButton } from "../Buttons";
+
+const HeaderButton = styled(PrimaryButton)`
+  font-size: 1.25rem;
+  padding: ${spacing[4]};
+  margin: 0;
+  margin-bottom: 10px;
+  margin-right: 10px;
+  transition: all 0.2s ease;
+  :hover {
+    background: ${colors.red[800]};
+  }
+`;
 
 const HeaderWrapper = styled.div`
   margin-top: 0px;
+  box-shadow: ${boxShadow.md};
   margin-bottom: 10px;
   display: flex;
+  background-size: cover;
   background-color: white;
   flex-direction: column;
   background: white;
@@ -15,61 +30,12 @@ const HeaderWrapper = styled.div`
   align-content: center;
   align-items: center;
   font-family: ${(props) => props.theme.MainFont1};
-  a {
-    align-self: flex-end;
-    justify-self: flex-end;
-    text-decoration: none;
-    text-align: center;
-    background-color: ${(props) => props.theme.mainColor1};
-    border-radius: 4px;
-    border: 1px solid #ccc;
-    color: white;
-    font-family: ${(props) => props.theme.MainFont1};
-    text-transform: uppercase;
-    transform-style: flat;
-    transition: all ease 0.3s;
-    letter-spacing: 0.18rem;
-    font-size: 0.8rem;
-    /* width: 120px; */
-    padding: 10px;
-    &:active {
-      box-shadow: 0 3px 0 #ccc;
-      top: 3px;
-      outline: none;
-    }
-    &:hover:enabled {
-      background-color: ${(props) => props.theme.mainColor2};
-      color: white;
-      cursor: pointer !important;
-    }
-    &:active:enabled {
-      background: ${(props) => props.theme.mainColor1} !important;
-      box-shadow: inset 0px 0px 5px #c1c1c1 !important;
-      outline: none;
-    }
-    @media (min-width: 360px) {
-      width: 130px;
-    }
-    @media (orientation: landscape) and (max-width: 568px) {
-      width: 160px;
-    }
-    @media (orientation: landscape) and (min-width: 569px) {
-      width: 180px;
-      padding: 12px;
-      margin-right: 20px;
-      margin-bottom: 8px;
-    }
-    /* @media screen and (max-width: 1024px) {
-      margin-top: 10px;
-    } */
-    @media screen and (max-width: 600px) {
-      align-self: center;
-      margin-top: 0px;
-    }
-  }
+
   .heading__content {
     display: flex;
-
+    /* new stuff */
+    align-items: center;
+    /* new Stuff */
     flex-direction: column;
     justify-content: space-evenly;
     height: 100%;
@@ -98,29 +64,57 @@ const HeaderWrapper = styled.div`
     font-family: ${(props) => props.theme.MainFont3};
     padding-right: 10px;
     padding-left: 10px;
+    font-weight: 700;
   }
   @media screen and (min-width: 360px) {
+    background: linear-gradient(
+        rgba(247, 250, 252, 0.85),
+        rgba(247, 250, 252, 0.45)
+      ),
+      url(${(props) => props.headerBG}) no-repeat;
+    background-size: cover;
     h2 {
       font-size: 1.65rem;
     }
   }
-  @media screen and (min-width: 730px) {
-    background: url(${(props) => props.headerBG}) no-repeat;
+  @media screen and (min-width: 760px) {
+    background: linear-gradient(
+        rgba(247, 250, 252, 0.45),
+        rgba(247, 250, 252, 0.45)
+      ),
+      url(${(props) => props.headerBG}) no-repeat;
+
     height: 200px;
     background-color: white;
     flex-direction: row;
     margin-bottom: 0px;
     .heading__content {
-      max-width: 350px;
+      align-items: flex-end;
+      max-width: 50%;
     }
-    a {
-      margin-right: 20px;
-      font-weight: 700;
-    }
+  }
+  @media screen and (min-width: 820px) {
+    background: url(${(props) => props.headerBG}) no-repeat;
+    background-size: cover;
   }
   @media screen and (min-width: 1000px) {
     .heading__content {
-      max-width: 550px;
+      align-items: flex-end;
+      max-width: 50%;
+    }
+    /* h2 {
+      font-size: 1.7rem;
+    }
+    p {
+      font-weight: 300;
+
+      font-size: 1.2rem;
+      line-height: 1.5rem;
+    } */
+  }
+  @media screen and (min-width: 1300px) {
+    .heading__content {
+      max-width: 50%;
     }
     h2 {
       font-size: 1.7rem;
@@ -132,21 +126,18 @@ const HeaderWrapper = styled.div`
       line-height: 1.5rem;
     }
   }
-  @media screen and (min-width: 1300px) {
-    .heading__content {
-      max-width: 650px;
-    }
-  }
 `;
-const Header = ({ description, title, headerBG, landing, brandName }) => {
+const Header = ({ description, title, headerBG, button }) => {
   return (
     <HeaderWrapper headerBG={headerBG}>
+      {/* {console.log(description, title, headerBG, landing, brandName, button)} */}
       <div className="heading__content">
         <div>
           <h2>{title}</h2>
           <p>{description}</p>
         </div>
-        {landing && <Link to={`/brands/${brandName}/landing`}>Learn more</Link>}
+        {/* {landing && <Link to={`/brands/${brandName}/landing`}>Learn more</Link>} */}
+        {button && <HeaderButton to={button.url}>{button.label}</HeaderButton>}
       </div>
     </HeaderWrapper>
   );
@@ -155,14 +146,12 @@ Header.propTypes = {
   description: PropTypes.string.isRequired,
   title: PropTypes.string.isRequired,
   headerBG: PropTypes.string,
-  landing: PropTypes.bool,
-  brandName: PropTypes.string,
+  button: PropTypes.shape({ label: PropTypes.string, url: PropTypes.string }),
 };
 
 Header.defaultProps = {
   headerBG: "",
-  landing: false,
-  brandName: "Test",
+  button: null,
 };
 
 export default Header;