+++ src/pages/brands.js
@@ -6,6 +6,7 @@ import { HelmetDatoCms } from "gatsby-source-datocms";
 import Layout from "../components/Layout";
 import LifeStyleCard from "../components/Brands/LifeStyleCard";
 import BreadCrumbs, { BreadWrapper } from "../components/BreadCrumbs";
+import Header from "../components/shared/ProductList/Header";
 
 const SectionContainer = styled.div`
   /* min-height: 100vh; */
@@ -26,6 +27,7 @@ const SectionContainer = styled.div`
   justify-items: center;
   justify-content: center;
   align-content: flex-start;
+  padding-top: 10px;
   img {
     max-width: 100%;
     height: auto;
@@ -39,8 +41,15 @@ const SectionContainer = styled.div`
     padding-right: 5px;
   }
   @media screen and (min-width: 1022px) {
-    padding-top: 40px;
+    padding-top: 20px;
     padding-bottom: 40px;
+    padding-left: 0px;
+    padding-right: 0px;
+  }
+`;
+const BrandsRoot = styled.div`
+  @media (min-width: 1022px) {
+    padding-top: 20px;
   }
 `;
 const Brands = ({ data }) => {
@@ -50,21 +59,29 @@ const Brands = ({ data }) => {
       <BreadWrapper hidenLarge>
         <BreadCrumbs here="Brands" />
       </BreadWrapper>
-      <SectionContainer>
-        {data.allDatoCmsBrand.nodes.map((brand) => (
-          <LifeStyleCard
-            key={brand.id}
-            mobileHeight="228px"
-            height="228px"
-            title={brand.displayName}
-            bgImg={brand.lifeStyleImg}
-            logo={brand.brandLogo}
-            url={`/brands/${brand.urlName}`}
-          >
-            {brand.lifeStyleText}
-          </LifeStyleCard>
-        ))}
-      </SectionContainer>
+      <BrandsRoot>
+        <Header
+          description={data.datoCmsHeader.tagLine}
+          headerBG={data.datoCmsHeader.bgImg.url}
+          button={{ label: "Shop all Mattresses", url: "/brands/list" }}
+          title="Our Brands"
+        />
+        <SectionContainer>
+          {data.allDatoCmsBrand.nodes.map((brand) => (
+            <LifeStyleCard
+              key={brand.id}
+              mobileHeight="228px"
+              height="228px"
+              title={brand.displayName}
+              bgImg={brand.lifeStyleImg}
+              logo={brand.brandLogo}
+              url={`/brands/${brand.urlName}`}
+            >
+              {brand.lifeStyleText}
+            </LifeStyleCard>
+          ))}
+        </SectionContainer>
+      </BrandsRoot>
       <BreadWrapper hidenLarge Bottom>
         <BreadCrumbs here="Brands" />
       </BreadWrapper>
@@ -77,6 +94,15 @@ export const brandsSEO = graphql`
     datoCmsSeo(name: { eq: "brands" }) {
       seoMetaTags {
         ...GatsbyDatoCmsSeoMetaTags
+      }
+    }
+    datoCmsHeader(title: { eq: "AllMattressSort" }) {
+      title
+      tagLine
+      bgImg {
+        title
+        alt
+        url
       }
     }
     allDatoCmsBrand(sort: { fields: position }) {