+++ src/pages/brands/sealy/index.jsx
@@ -29,7 +29,10 @@ const Sealy = ({ data }) => {
         description={datoCmsBrand.headerLink.tagLine}
         breadCrumbs
         brandName={datoCmsBrand.urlName}
-        landing
+        button={{
+          label: "Learn More",
+          url: `/brands/${datoCmsBrand.urlName}/landing`,
+        }}
       />
     </Layout>
   );